refactor(Integrations/WPML) filter matchers, remove Filters class in favour of Rewrite
