+++ src/Tribe/Integrations/WPML/Rewrites.php
@@ -1,5 +1,8 @@
 <?php
 
+use Tribe\Events\I18n;
+use Tribe__Rewrite as Common_Rewrite;
+
 
 /**
  * Class Tribe__Events__Integrations__WPML__Rewrites
@@ -42,6 +45,15 @@ class Tribe__Events__Integrations__WPML__Rewrites {
 	 * @var array
 	 */
 	protected $replacement_rules = [];
+
+	/**
+	 * A map from language codes to the set of translated bases.
+	 *
+	 * @since TBD
+	 *
+	 * @var array<string,array<string,string>>
+	 */
+	private array $bases_by_language = [];
 
 	/**
 	 * @return Tribe__Events__Integrations__WPML__Linked_Posts
@@ -92,7 +104,7 @@ class Tribe__Events__Integrations__WPML__Rewrites {
 	}
 
 	protected function prepare_venue_slug_translations() {
-		$wpml_i18n_strings = Tribe__Events__Integrations__WPML__Utils::get_wpml_i18n_strings(
+		$wpml_i18n_strings             = Tribe__Events__Integrations__WPML__Utils::get_wpml_i18n_strings(
 			[ $this->venue_slug ]
 		);
 		$post_slug_translations        = Tribe__Events__Integrations__WPML__Utils::get_post_slug_translations_for( Tribe__Events__Venue::POSTTYPE );
@@ -207,9 +219,188 @@ class Tribe__Events__Integrations__WPML__Rewrites {
 
 		// Fetch translated versions of the event category slug and append them
 		$category_translation = Tribe__Events__Integrations__WPML__Category_Translation::instance();
-		$translated_slugs = $category_translation->get_translated_base_slugs();
-		$bases['tax'] = array_merge( $bases['tax'], $translated_slugs );
+		$translated_slugs     = $category_translation->get_translated_base_slugs();
+		$bases['tax']         = array_merge( $bases['tax'], $translated_slugs );
 
 		return $bases;
+	}
+
+	/**
+	 * Translate the Event single slugs.
+	 *
+	 * @param array<string,array<string>> $bases The bases to translate.
+	 *
+	 * @return array<string,array<string>> The translated bases.
+	 */
+	protected function translate_single_slugs( array $bases ) {
+		global $sitepress_settings;
+
+		$supported_post_types = [ Tribe__Events__Main::POSTTYPE ];
+
+		foreach ( $supported_post_types as $post_type ) {
+			// check that translations are active for this CPT
+			$cpt_slug_is_not_translated = empty( $sitepress_settings['posts_slug_translation']['types'][ $post_type ] );
+
+			if ( $cpt_slug_is_not_translated ) {
+				continue;
+			}
+
+			$event_slug = WPML_Slug_Translation::get_slug_by_type( $post_type );
+
+			$string_id = icl_get_string_id( $event_slug, 'WordPress', 'URL slug: ' . $post_type );
+
+			if ( ! $string_id ) {
+				continue;
+			}
+
+			$slug_translations = icl_get_string_translations_by_id( $string_id );
+
+			if ( empty( $slug_translations ) ) {
+				continue;
+			}
+
+			$bases['single'] = array_merge( $bases['single'], wp_list_pluck( $slug_translations, 'value' ) );
+		}
+
+		return $bases;
+	}
+
+	/**
+	 * Translate the Event archive slugs.
+	 *
+	 * @since TBD
+	 *
+	 * @param array<string,array<string>> $bases The bases to translate.
+	 *
+	 *
+	 * @return array<string,array<string>> The translated bases.
+	 */
+	protected function translate_archive_slugs( array $bases ) {
+		$supported_post_types = array( Tribe__Events__Main::POSTTYPE );
+
+		foreach ( $supported_post_types as $post_type ) {
+
+			$slug = Tribe__Settings_Manager::get_option( 'eventsSlug', 'events' );
+
+			$context   = [ 'domain' => 'the-events-calendar', 'context' => 'Archive Events Slug' ];
+			$string_id = icl_get_string_id( $slug, $context );
+
+			if ( ! $string_id ) {
+				// If we couldn't find the string, we might need to register it.
+				icl_register_string( $context, false, $slug );
+
+				continue;
+			}
+
+			$slug_translations = icl_get_string_translations_by_id( $string_id );
+
+			if ( empty( $slug_translations ) ) {
+				continue;
+			}
+
+			$bases['archive'] = array_merge( $bases['archive'], wp_list_pluck( $slug_translations, 'value' ) );
+		}
+
+		return $bases;
+	}
+
+	/**
+	 * Filters the bases used to generate TEC rewrite rules to use WPML managed translations.
+	 *
+	 * @param array  $bases
+	 * @param string $method
+	 * @param array  $domains
+	 *
+	 * @return array An array of bases each with its (optional) WPML managed translations set.
+	 */
+	public function filter_tribe_events_rewrite_i18n_slugs_raw( $bases, $method, $domains ) {
+		/** @var SitePress $sitepress */
+		global $sitepress, $sitepress_settings;
+
+		if ( empty( $sitepress ) || ! $sitepress instanceof SitePress ) {
+			return $bases;
+		}
+
+		// Grab all languages
+		$langs = $sitepress->get_active_languages();
+
+		// Sort the languages to stick w/ the order that will be used to support localized bases.
+		ksort( $langs );
+
+		if ( empty( $langs ) ) {
+			return $bases;
+		}
+
+		foreach ( $langs as $lang ) {
+			$languages[] = $sitepress->get_locale( $lang['code'] );
+		}
+
+		// Prevent Duplicates and Empty langs
+		$languages = array_filter( array_unique( $languages ) );
+
+		// Query the Current Language
+		$current_locale = $sitepress->get_locale( $sitepress->get_current_language() );
+
+		// Get the strings on multiple Domains and Languages
+		// remove WPML filter to avoid the locale being set to the default one
+		remove_filter( 'locale', [ $sitepress, 'locale_filter' ] );
+
+		/*
+		 * Translate only the English version of the bases to ensure the order of the translations.
+		 */
+		$untranslated_bases = array_combine( array_keys( $bases ), array_column( $bases, 0 ) );
+
+		$i18n        = tribe( 'tec.i18n' );
+		$flags       = I18n::COMPILE_STRTOLOWER | I18n::RETURN_BY_LANGUAGE;
+		$by_language = $i18n->get_i18n_strings( $untranslated_bases, $languages, $domains, $current_locale, $flags );
+		// Store this value to use it in the `filter_localized_matchers` method.
+		$this->bases_by_language = $by_language;
+
+		// Merge and deduplicate; the `get_i18n_strings` would do this, but the language information would be lost.
+		$translated_bases = array_merge_recursive( ...array_values( $by_language ) );
+		foreach ( $translated_bases as &$set ) {
+			$set = array_unique( $set );
+		}
+		unset( $set );
+
+		// Prepend the WPML-translated bases to the set of bases.
+		$bases = array_merge_recursive( $translated_bases, $bases );
+
+		// re-hook WPML filter
+		add_filter( 'locale', [ $sitepress, 'locale_filter' ] );
+
+		$string_translation_active = defined( 'WPML_ST_VERSION' );
+		$post_slug_translation_on  = ! empty( $sitepress_settings['posts_slug_translation']['on'] );
+
+		if ( $string_translation_active && $post_slug_translation_on ) {
+			$bases = $this->translate_single_slugs( $bases );
+			$bases = $this->translate_archive_slugs( $bases );
+		}
+
+		return $bases;
+	}
+
+	/**
+	 * Filters the localized matcher to use WPML managed translations.
+	 *
+	 * @since TBD
+	 *
+	 * @param string $localized_slug The matcher localized slug.
+	 * @param string $query_var      The query var the matcher is for.
+	 *
+	 * @return string The localized slug.
+	 */
+	public function localize_matcher( $localized_slug, $query_var ) {
+		if ( ! ( is_string( $localized_slug ) && is_string( $query_var ) ) ) {
+			return $localized_slug;
+		}
+
+		$current_language = get_locale();
+
+		if ( ! empty( $this->bases_by_language[ $current_language ][ $query_var ] ) ) {
+			return end( $this->bases_by_language[ $current_language ][ $query_var ] );
+		}
+
+		return $localized_slug;
 	}
 }
